// src/app/api/bakery/path.ts
export enum PATHS {
  ALL_BAKERY = "/driver-bread/get-all-baker-rooms",
  BAKERY = "/driver-bread/get-baker-room/",
  BAKERY_DOUGH_START = "/driver-bread/",
  BAKERY_DOUGH_INFO_END = "/give-dough-info",
  BAKERY_DOUGH_END = "/give-dough",
  BAKERY_DELIVERY_END = "/yetkazib-beruvchi",
  BAKERY_BREAD_INFO_END = "/give-take-bread-info",
  BAKERY_BREAD_END = "/give-take-bread",
}